[Cleanup] fix broken sovryn logo usage, consolidate babelfish icon usage
